stub in-memory storage/blob implementation
